Fix bug and refactor MaximalBlockBuilder.build()
m_end_addr should have the value addr(last_inst) + size(last_inst)
when checking overlap.

+++ src/disasm/BasicBlock.cpp
@@ -14,7 +14,7 @@
 
 namespace disasm {
 
-BasicBlock::BasicBlock(unsigned id, const cs_insn *inst) :
+BasicBlock::BasicBlock(size_t id, const cs_insn *inst) :
     m_valid{false},
     m_id{id},
     m_size{0} {
