+++ src/disasm/BasicBlock.h
@@ -25,7 +25,7 @@ public:
      * methods other than operator= and valid on this results in
      * undefined behavior.
      */
-    BasicBlock(unsigned id, const cs_insn *inst);
+    BasicBlock(size_t id, const cs_insn *inst);
     virtual ~BasicBlock() = default;
     BasicBlock(const BasicBlock &src) = default;
     BasicBlock &operator=(const BasicBlock &src) = default;
@@ -50,7 +50,7 @@ private:
 
 private:
     bool m_valid;
-    unsigned m_id;
+    size_t m_id;
     size_t m_size;
     std::vector<addr_t> m_inst_addrs;
 };
